feat: 기존 review, 사용자 리뷰  반영

+++ back/src/main/java/com/a505/jupasu/domain/reviews/dto/MyPageReviewResponse.java
@@ -20,7 +20,7 @@ public class MyPageReviewResponse {
     private String nickname;
     private Long wineId;
     private String wineName;
-    private Float rating;
+    private Integer rating;
     private String content;
     private LocalDateTime createdAt;
